添加分析数据
