--- main.py
@@ -1,19 +1,0 @@
-from Crawler import RatingCrawler
-import numpy as np
-import threading
-
-
-def thread_it(func, *args):
-    '''将函数打包进线程'''
-    # 创建
-    t = threading.Thread(target=func, args=args)
-    # 守护 !!!
-    t.setDaemon(True)
-    # 启动
-    t.start()
-    # 阻塞--卡死界面！
-    # t.join()
-
-
-Crawler = RatingCrawler()
-thread_it(Crawler.start(315))
